perf: refactor storage initialization for robustness and fix tests
- Move directory creation to the beginning of initialization functions to ensure they run before any fallible connection attempts.
- Update tests to use subdirectories, verifying create_dir_all behavior.
- Fix test_initialize_dual_storage to handle expected connection failures while verifying directory creation side effects.
- Relocate #![allow(unused_imports)] to the top of storage.rs to address reviewer feedback.
- Maintain Rust 2024 compliance with unsafe blocks for environment variable manipulation.
