+++ memory-storage-turso/src/storage/episodes/query.rs
@@ -57,13 +57,10 @@ impl TursoStorage {
 
         // Apply limit with defaults and bounds
         let limit = apply_query_limit(query.limit);
-        sql.push_str(" LIMIT ?");
-
-        let mut params: Vec<libsql::Value> = params_vec.into_iter().map(|p| p.into()).collect();
-        params.push((limit as i64).into());
+        sql.push_str(&format!(" LIMIT {}", limit));
 
         let mut rows = conn
-            .query(&sql, libsql::params_from_iter(params))
+            .query(&sql, libsql::params_from_iter(params_vec))
             .await
             .map_err(|e| Error::Storage(format!("Failed to query episodes: {}", e)))?;
