+++ memory-storage-turso/src/storage/tag_operations.rs
@@ -4,13 +4,9 @@
 //! tag-based queries, and tag statistics.
 
 use crate::{Error, Result};
-use do_memory_core::apply_query_limit;
 use libsql::Connection;
 use std::collections::HashMap;
 use uuid::Uuid;
-
-/// Maximum number of tags allowed in a single query to prevent performance issues
-const MAX_TAGS_PER_QUERY: usize = 100;
 
 /// Tag statistics
 #[derive(Debug, Clone)]
@@ -124,13 +120,6 @@ pub async fn delete_episode_tags(
 
     let episode_id_str = episode_id.to_string();
 
-    // Apply bound to tags
-    let tags = if tags.len() > MAX_TAGS_PER_QUERY {
-        &tags[..MAX_TAGS_PER_QUERY]
-    } else {
-        tags
-    };
-
     // Build placeholders for IN clause
     let placeholders = tags.iter().map(|_| "?").collect::<Vec<_>>().join(",");
     let query = format!(
@@ -161,22 +150,15 @@ pub async fn find_episodes_by_tags_or(
         return Ok(Vec::new());
     }
 
-    let tags = if tags.len() > MAX_TAGS_PER_QUERY {
-        &tags[..MAX_TAGS_PER_QUERY]
-    } else {
-        tags
-    };
-
     let placeholders = tags.iter().map(|_| "?").collect::<Vec<_>>().join(",");
-    let effective_limit = apply_query_limit(limit);
+    let limit_clause = limit.map(|l| format!(" LIMIT {}", l)).unwrap_or_default();
 
     let query = format!(
-        "SELECT DISTINCT episode_id FROM episode_tags WHERE tag IN ({}) ORDER BY created_at DESC LIMIT ?",
-        placeholders
+        "SELECT DISTINCT episode_id FROM episode_tags WHERE tag IN ({}) ORDER BY created_at DESC{}",
+        placeholders, limit_clause
     );
 
-    let mut params: Vec<libsql::Value> = tags.iter().map(|t| t.clone().into()).collect();
-    params.push((effective_limit as i64).into());
+    let params: Vec<libsql::Value> = tags.iter().map(|t| t.clone().into()).collect();
 
     let stmt = conn
         .prepare(&query)
@@ -215,15 +197,9 @@ pub async fn find_episodes_by_tags_and(
         return Ok(Vec::new());
     }
 
-    let tags = if tags.len() > MAX_TAGS_PER_QUERY {
-        &tags[..MAX_TAGS_PER_QUERY]
-    } else {
-        tags
-    };
-
     let tag_count = tags.len();
     let placeholders = tags.iter().map(|_| "?").collect::<Vec<_>>().join(",");
-    let effective_limit = apply_query_limit(limit);
+    let limit_clause = limit.map(|l| format!(" LIMIT {}", l)).unwrap_or_default();
 
     // Query: Find episodes that have all specified tags
     let query = format!(
@@ -234,14 +210,13 @@ pub async fn find_episodes_by_tags_and(
         GROUP BY episode_id
         HAVING COUNT(DISTINCT tag) = ?
         ORDER BY MAX(created_at) DESC
-        LIMIT ?
+        {}
         "#,
-        placeholders
+        placeholders, limit_clause
     );
 
     let mut params: Vec<libsql::Value> = tags.iter().map(|t| t.clone().into()).collect();
     params.push((tag_count as i64).into());
-    params.push((effective_limit as i64).into());
 
     let stmt = conn
         .prepare(&query)
