+++ scripts/diagnose-mcp.sh
