+++ scripts/fix-file-locations.sh
