+++ scripts/setup-target-dir.sh
