+++ scripts/test-mcp-tools.sh
