+++ scripts/test-memory-mcp.sh
