+++ scripts/test_advanced_patterns.sh
