+++ scripts/test_mcp_comprehensive.sh
