+++ scripts/test_mcp_sequence.sh
