+++ scripts/test_memory_mcp_prompt.sh
