+++ scripts/validate_staging_optimizations.sh
