+++ scripts/verify_cli_usage.sh
