+++ scripts/verify_embeddings_integration.sh
